+++ Assets/Project/Scripts/GamePlayScene/Bullet/NormalCartridgeController.cs
@@ -1,5 +1,5 @@
 using System;
-using Project.Scripts.Utils.Attribute;
+using Project.Scripts.Utils.Attributes;
 using Project.Scripts.Utils.Definitions;
 using Project.Scripts.Utils.Library.Extension;
 using UnityEngine;
