+++ Assets/Project/Scripts/Utils/Attributes/NonEditableAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace Project.Scripts.Utils.Attribute
+namespace Project.Scripts.Utils.Attributes
 {
     public class NonEditableAttribute : PropertyAttribute {}
 }
